Refactoring how the data context is initialized and how it is seeded.  Still need to update most tests.  Checking in now to get baseline build before changes.
