+++ src/Microsoft.Restier.Tests.AspNet/Microsoft.Restier.Tests.AspNet.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net472</TargetFramework>
+    <DefineConstants>$(DefineConstants);EF6</DefineConstants>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
