+++ src/Microsoft.Restier.Tests.AspNetCorePlusEF6/UnitTest1.cs
@@ -1,0 +1,13 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Microsoft.Restier.Tests.AspNetCorePlusEF6
+{
+    [TestClass]
+    public class UnitTest1
+    {
+        [TestMethod]
+        public void TestMethod1()
+        {
+        }
+    }
+}
