+++ src/Microsoft.Restier.Tests.Shared.EntityFramework/Scenarios/Library/LibraryContext.cs
@@ -1,22 +1,23 @@
 ﻿// Copyright (c) Microsoft Corporation.  All rights reserved.
 // Licensed under the MIT License.  See License.txt in the project root for license information.
 
-using System;
-using System.Collections.ObjectModel;
+#if EF6
 using System.Data.Entity;
-using Microsoft.OData.Edm;
+#else
+using Microsoft.EntityFrameworkCore;
+#endif
 
 namespace Microsoft.Restier.Tests.Shared.Scenarios.Library
 {
 
     /// <summary>
-    /// The EntityFramework data context for the Library scenario.
+    /// The data context for the Library scenario.
     /// </summary>
     public class LibraryContext : DbContext
     {
 
-        public LibraryContext()
-            : base("LibraryContext") => Database.SetInitializer(new LibraryTestInitializer());
+#if EF6
+        #region EntitySet Properties
 
         public IDbSet<Book> Books { get; set; }
 
@@ -30,49 +31,39 @@ namespace Microsoft.Restier.Tests.Shared.Scenarios.Library
 
         public IDbSet<Universe> Universes { get; set; }
 
-    }
+        #endregion
 
-    /// <summary>
-    /// An initializer to populate data into the context.
-    /// </summary>
-    public class LibraryTestInitializer : DropCreateDatabaseAlways<LibraryContext>
-    {
-        protected override void Seed(LibraryContext context)
+        public LibraryContext()
+            : base("LibraryContext") => Database.SetInitializer(new LibraryTestInitializer());
+
+#else
+
+        #region EntitySet Properties
+
+        public DbSet<Book> Books { get; set; }
+
+        public DbSet<LibraryCard> LibraryCards { get; set; }
+
+        public DbSet<Publisher> Publishers { get; set; }
+
+        public DbSet<Employee> Readers { get; set; }
+
+        public DbSet<Address> Addresses { get; set; }
+
+        public DbSet<Universe> Universes { get; set; }
+
+        #endregion
+
+        public LibraryContext(DbContextOptions<LibraryContext> options)
+        : base(options)
+        { }
+
+        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
         {
-            var sourceData = new LibraryTestDataFactory();
-
-            if (sourceData.Addresses is not null)
-            {
-                sourceData.Addresses.ForEach(c => context.Addresses.Add(c));
-            }
-
-            if (sourceData.Universes is not null)
-            {
-                sourceData.Universes.ForEach(c => context.Universes.Add(c));
-            }
-
-            if (sourceData.Books is not null)
-            {
-                sourceData.Books.ForEach(c => context.Books.Add(c));
-            }
-
-            if (sourceData.LibraryCards is not null)
-            {
-                sourceData.LibraryCards.ForEach(c => context.LibraryCards.Add(c));
-            }
-
-            if (sourceData.Publishers is not null)
-            {
-                sourceData.Publishers.ForEach(c => context.Publishers.Add(c));
-            }
-
-            if (sourceData.Readers is not null)
-            {
-                sourceData.Readers.ForEach(c => context.Readers.Add(c));
-            }
-
-            context.SaveChanges();
+            optionsBuilder.UseInMemoryDatabase(nameof(LibraryContext));
         }
 
+#endif
     }
+
 }
