+++ src/Microsoft.Restier.Tests.Shared.EntityFramework/Scenarios/Marvel/MarvelApi.cs
@@ -2,14 +2,23 @@
 // Licensed under the MIT License.  See License.txt in the project root for license information.
 
 using System;
-using Microsoft.Restier.EntityFramework;
+#if NET5_0_OR_GREATER
+    using Microsoft.Restier.AspNetCore.Model;
+#else
+    using Microsoft.Restier.AspNet.Model;
+#endif
+
+#if EF6
+    using Microsoft.Restier.EntityFramework;
+#else
+    using Microsoft.Restier.EntityFrameworkCore;
+#endif
 
 namespace Microsoft.Restier.Tests.Shared.Scenarios.Marvel
 {
 
     /// <summary>
     /// A testable API that implements an Entity Framework model and has secondary operations
-    /// against a SQL 2017 LocalDB database.
     /// </summary>
     public class MarvelApi : EntityFrameworkApi<MarvelContext>
     {
